Smart_SDK _2/8

+++ src/main/java/com/market/alphavantage/entity/StockPrice.java
@@ -25,6 +25,6 @@ public class StockPrice {
 
     private Double[] close;
 
-    private Long[] volume;
+    private Double[] volume;
 }
